% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{add_comments}
\alias{add_comments}
\title{Extend code regions to include leading comments and whitespace}
\usage{
add_comments(refs)
}
\arguments{
\item{refs}{a list of the code region \code{srcref}s to extend.}
}
\value{
\code{add_comment} returns a list of \code{srcref}s corresponding to
\code{srcref}, but extended to include the preceding comment block.
}
\description{
Extend code regions to include leading comments and whitespace
}
\keyword{internal}
