% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{attach_module}
\alias{attach_module}
\title{Attach a module environment locally or globally}
\usage{
attach_module(all, operators, name, mod_env, parent)
}
\arguments{
\item{all}{logical specifying whether to attach the whole module}

\item{operators}{logical specifying whether to attach operators}

\item{name}{the module name}

\item{mod_env}{the module environment to attach}

\item{parent}{the module parent environment}
}
\description{
Attach a module environment locally or globally
}
\details{
If neither \code{all} nor \code{operators} are \code{TRUE}, this function
does nothing. Otherwise, it will either attach the whole module environment
or the operators it exports. Attaching is done by inserting the module at the
first position into the parent environment chain.
}

