% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-package.r
\docType{package}
\name{box}
\alias{box}
\alias{box-package}
\title{An alternative module system for R}
\description{
Use \code{box::use(prefix/mod)} to import a module, or \code{box::use(pkg)}
to import a package. Fully qualified names are supported for nested modules,
reminiscent of module systems in many other modern languages.
}
\section{Using modules & packages}{


\itemize{
 \item \code{\link{use}}
}
}

\section{Writing modules}{


Infrastructure and utility functions that are mainly used inside modules.

\itemize{
 \item \code{\link{file}}
 \item \code{\link{name}}
 \item \code{\link{register_S3_method}}
 \item \link{mod-hooks}
}
}

\section{Interactive use}{


Functions for use in interactive sessions and for testing.

\itemize{
 \item \code{\link{help}}
 \item \code{\link{unload}}, \code{\link{reload}}
 \item \code{\link{set_script_path}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://klmr.me/box/}
  \item \url{https://github.com/klmr/box}
  \item Report bugs at \url{https://github.com/klmr/box/issues}
}

}
\author{
\strong{Maintainer}: Konrad Rudolph \email{konrad.rudolph@gmail.com} (\href{https://orcid.org/0000-0002-9866-7051}{ORCID})

Other contributors:
\itemize{
  \item Michael Schubert \email{mschu.dev@gmail.com} (\href{https://orcid.org/0000-0002-6862-5221}{ORCID}) [contributor]
}

}
