% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/access.r
\name{$.module}
\alias{$.module}
\title{Access an object inside a module}
\usage{
module$object
}
\arguments{
\item{module}{the module object}

\item{object}{the name of the object}
}
\value{
Unlike the default \code{$} operator in R, access non-existent
objects in modules yields an error rather than returning \code{NULL}.
}
\description{
Access an object inside a module
}

