% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{create_export_block}
\alias{create_export_block}
\alias{parse_object}
\alias{roxygen2_object}
\title{Collect export tag information}
\usage{
create_export_block(expr, ref, info, mod_ns)

parse_object(info, expr, mod_ns)

roxygen2_object(alias, value, type)
}
\arguments{
\item{expr}{The unevaluated expression represented by the tag.}

\item{ref}{The code reference \code{srcref} represented by the tag.}

\item{alias}{The object name.}

\item{value}{The object value.}

\item{type}{The object type.}
}
\value{
\code{create_export_block} returns an object of type
\code{roxy_block} represents an exported declaration expression, along with
its source code location.
}
\description{
Collect export tag information
}
\note{
This could be represented much simpler but we keep compatibility with
roxygen2 — at least for the time being — to make integration with the
roxygen2 API easier, should it become necessary.
}
\keyword{internal}
