% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{do_import}
\alias{do_import}
\title{Perform the actual import operation on an individual module}
\usage{
do_import(module_name, module_path, doc)
}
\arguments{
\item{module_name}{the name of the module, as specified by the user}

\item{module_path}{the fully resolved path to the module file}

\item{doc}{logical, whether the module documentation should be parsed}
}
\description{
Perform the actual import operation on an individual module
}

