% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhibit_namespace.r
\name{exhibit_namespace}
\alias{exhibit_module_namespace}
\alias{exhibit_namespace}
\alias{exhibit_package_namespace}
\title{Exhibit objects from an internal module namespace as an environment}
\usage{
exhibit_namespace(objects, name, path, doc, parent)

exhibit_module_namespace(namespace, name, parent, export_list)

exhibit_package_namespace(namespace, name, parent, export_list)
}
\arguments{
\item{objects}{named list of objects}

\item{name}{the name of the resulting environment}

\item{path}{the fully resolved path to the corresponding module or package}

\item{doc}{the module documentation}

\item{parent}{the parent environment of the resulting environment}

\item{namespace}{the namespace to export object from}

\item{export_list}{the list of objects to export; if \code{NULL}, export
everything}
}
\value{
Returns the resulting environment.
}
\description{
Exhibit objects from an internal module namespace as an environment

\code{exhibit_module_namespace} exports a namespace for a module.

\code{exhibit_package_namespace} exports a namespace for a package.
}

