% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{export_operators}
\alias{export_operators}
\title{Copy a module’s operators into a separate environment}
\usage{
export_operators(environment, parent, module_name)
}
\arguments{
\item{environment}{the module environment}

\item{parent}{the parent environment of the calling code, to determine how to
chain the environments properly}

\item{module_name}{the name of the module}
}
\value{
A new environment containing the operators of the module.
}
\description{
This function is used to create an attachable environment containing only the
module’s operators, as these would otherwise not be readily usable.
}
\note{
This function expects the module \emph{environment} rather than the
\emph{namespace}. This is important because, for package modules, we only
want to expose exported operators, and not all exported operators are visible
inside the namespace (see, for example, \code{\link[dplyr]{\%>\%}}, which is
imported from the package “magrittr”).
}

