% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{fix_module_attributes}
\alias{fix_module_attributes}
\title{Save a module’s inherited attributes into a local environment}
\usage{
fix_module_attributes(envir)
}
\arguments{
\item{envir}{the environment into which to copy the inherited attributes}
}
\description{
Save a module’s inherited attributes into a local environment
}
\details{
Attaching a module to an environment makes the environment “forget” its
actual enclosing environment. This is important since that module holds
information such as the module name or path, which may be required in the
current local scope. To fix this, \code{fix_module_attributes} copies the
enclosing module’s attributes into the local environment.
}

