% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{has_export_tag}
\alias{has_export_tag}
\title{Find \code{@export} tags in code regions}
\usage{
has_export_tag(ref)
}
\arguments{
\item{ref}{The code region \code{srcref} to search.}
}
\value{
\code{TRUE} if the given region is annotated with a \code{@export}
tag, \code{FALSE} otherwise.
}
\description{
Find \code{@export} tags in code regions
}
\keyword{internal}
