% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_module.r
\name{import_search_path}
\alias{import_search_path}
\title{Return the import module search path}
\usage{
import_search_path()
}
\description{
Return the import module search path
}
\note{
The search paths are ordered from highest to lowest priority.
The current module’s path always has the lowest priority.

There are two ways of modifying the module search path: by default,
\code{options('import.path')} specifies the search path. If and only if that
is unset, R also considers the environment variable \code{R_IMPORT_PATH}.
}

