% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{info}
\alias{info}
\alias{mod_info}
\alias{pkg_info}
\alias{crr_info}
\title{Information about a physical module or package}
\usage{
mod_info(spec, source_path)

pkg_info(spec)

crr_info(spec, inst_mod_path)
}
\arguments{
\item{spec}{a \code{mod_spec} or \code{pkg_spec} (for \code{mod_info} and
\code{pkg_info}, respectively)}

\item{source_path}{character string full path to the physical module
location.}
}
\value{
\code{mod_info} and \code{pkg_info} return a structure representing
the module/package information for the given specification/source location.
}
\description{
A \code{mod_info} represents an existing, installed module and its runtime
physical location (usually in the file system).
}
\keyword{internal}
