% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_S3_user_generic}
\alias{is_S3_user_generic}
\title{Check whether a function given by name is a user-defined generic}
\usage{
is_S3_user_generic(function_name, envir = parent.frame())
}
\arguments{
\item{function_name}{function name as character string}
}
\description{
A user-defined generic is any function which, at some point, calls
}

