% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cache.r
\docType{data}
\name{loaded_modules}
\alias{loaded_modules}
\title{Environment of loaded modules}
\format{\preformatted{<environment: 0x7fd17968b358> 
}}
\usage{
loaded_modules
}
\description{
Each module is stored as an environment inside \code{loaded_modules} with
the module’s code location path as its identifier. The path rather than the
module name is used because module names are not unique: two modules called
\code{a} can exist nested inside modules \code{b} and \code{c}, respectively.
Yet these may be loaded at the same time and need to be distinguished.
}
\keyword{datasets}

