% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.r
\name{lsf}
\alias{lsf}
\title{Return a list of function names in an environment}
\usage{
lsf(envir)
}
\arguments{
\item{envir}{the environment to search in.}
}
\value{
\code{lsf} returns a vector of function names in the given environment.
}
\description{
Return a list of function names in an environment
}
\keyword{internal}
