% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cache.r
\name{module_attributes}
\alias{module_attr}
\alias{module_attr<-}
\alias{module_attributes}
\alias{module_attributes<-}
\title{Module attributes}
\usage{
module_attributes(module)

module_attributes(module) <- value

module_attr(module, attr)

module_attr(module, attr) <- value
}
\arguments{
\item{module}{a module}

\item{value}{the attributes to assign}

\item{attr}{the attribute name}
}
\description{
\code{module_attributes} returns or assigns the attributes associated with
a module.

\code{module_attr} reads or assigns a single attribute associated with a
module.
}

