\name{module_base_path}
\alias{module_base_path}
\title{Get a module’s base directory}
\usage{
module_base_path(module)
}
\arguments{
  \item{module}{a module environment or namespace}
}
\value{
A character string containing the module’s base directory,
or the current working directory if not invoked on a
module.
}
\description{
Get a module’s base directory
}

