% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cache.r
\name{module_path}
\alias{module_path}
\title{Get a module’s path}
\usage{
module_path(module)
}
\arguments{
\item{module}{a module environment or namespace}
}
\value{
A character string containing the module’s full path.
}
\description{
Get a module’s path
}

