% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{parse_export_tags}
\alias{parse_export_tags}
\title{Extract comment tags from Roxygen block comments}
\usage{
parse_export_tags(info, exprs, mod_ns)
}
\arguments{
\item{exprs}{The unevaluated expressions to parse.}
}
\value{
\code{parse_export_tags} returns a list of \code{roxy_block}s for all
exported declarations.
}
\description{
Extract comment tags from Roxygen block comments
}
\note{
The following code performs the same function as roxygen2 with a custom
\code{@} tag roclet. Unfortunately roxygen2 itself pulls in many
dependencies, making it less suitable for an infrastructure package such as
this one. Furthermore, the code license of roxygen2 is incompatible with
ours, so we cannot simply copy and paste the relevant code out. Luckily the
logic is straightforward to reimplement.
}
\keyword{internal}
