% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.r
\name{rethrow}
\alias{rethrow}
\alias{rethrow_on_error}
\title{Pretty-print error call to be informative for the user.}
\usage{
rethrow(error, call = sys.call(sys.parent()))

rethrow_on_error(expr, call = sys.call(sys.parent()))
}
\arguments{
\item{error}{an object of class \code{c("error", "condition")} to rethrow}

\item{call}{the calling context to rethrow the error from, overwriting the
error’s current call object}

\item{expr}{an expression to evaluate inside \code{tryCatch}}
}
\value{
If it does not throw an error, \code{rethrow_on_error} returns the
value of evaluating \code{expr}.
}
\description{
Pretty-print error call to be informative for the user.
}
\keyword{internal}
