% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded.r
\name{set_script_path}
\alias{set_script_path}
\title{Set the base path of the script}
\usage{
set_script_path(path = NULL)
}
\arguments{
\item{path}{character string containing the relative or absolute path to the
currently executing R code file, or \code{NULL} to reset the path.}
}
\value{
\code{box::set_script_path} returns the previously set script path,
or \code{NULL} if none was explicitly set.
}
\description{
Set the base path of the script
}
\details{
\pkg{box} needs to know the base path of the topmost calling R context (i.e.
the script) to find relative import locations. In most cases, \code{box} can
figure the path out automatically. However, in some cases third-party
packages load code in a way in which \pkg{box} cannot find the correct path
of the script any more. \code{set_script_path} can be used in these cases to
set the path of the currently executing R script manually.
}
