% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_module.r
\name{split_path}
\alias{split_path}
\alias{merge_path}
\title{Split a path into its components and merge them back together}
\usage{
split_path(path)

merge_path(components)
}
\arguments{
\item{path}{the path to split}

\item{components}{character string vector of path components to merge}
}
\value{
\code{split_path} returns a character vector of path components that
logically represent \code{path}.

\code{merge_path} returns a single character string that is
logically equivalent to the \code{path} passed to \code{split_path}.
logically represent \code{path}.
}
\description{
\code{split_path(path)} is a platform independent and file system logic
aware alternative to \code{strsplit(path, '/')[[1]]}.
}
\note{
\code{merge_path} is the inverse function to \code{split_path}.
However, this does not mean that its result will be identical to the
original path. Instead, it is only guaranteed that it will refer to the same
logical path given the same working directory.
}
