% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.r
\name{wrap_unsafe_function}
\alias{wrap_unsafe_function}
\title{Wrap \dQuote{unsafe calls} functions}
\usage{
wrap_unsafe_function(ns, name)
}
\arguments{
\item{ns}{The namespace of the unsafe function.}

\item{name}{The name of the unsafe function.}
}
\value{
\code{wrap_unsafe_calls} returns a wrapper function with the same
argument as the wrapped function that can be called without causing a NOTE.
}
\description{
\code{wrap_unsafe_function} declares a function wrapper to a function that
causes an \command{R CMD check} NOTE when called directly. We should usually
not call these functions, but we need some of them because we want to
explicitly support features they provide.
}
\note{
Using an implementation that simply aliases \code{getExportedValue}
does not work, since \command{R CMD check} sees right through this
\dQuote{ruse}.
}
\keyword{internal}
