% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{find_in_path}
\alias{find_in_path}
\title{Find a module’s source location}
\usage{
find_in_path(spec, base_paths)
}
\arguments{
\item{spec}{a \code{mod_spec}.}

\item{base_paths}{a character vector of paths to search the module in, in
order of preference.}
}
\value{
\code{find_in_path} returns a \code{mod_info} that specifies the
module source location.
}
\description{
Find a module’s source location
}
\details{
A module is physically represented in the file system either by
\file{‹spec_name(spec)›.r} or by \file{‹spec_name(spec)›/__init__.r}, in that
order of preference in case both exist. File extensions are case insensitive
to allow for R’s obsession with capital-R extensions (but lower-case are
given preference, and upper-case file extensions are discouraged).
}
\keyword{internal}
