% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.r
\name{module_help}
\alias{?}
\alias{help}
\alias{module_help}
\title{Display module documentation}
\usage{
module_help(topic, help_type = getOption("help_type", "text"))

# ?module$function

# help(module$function)
}
\arguments{
\item{topic}{fully-qualified name of the object or function to get help for,
in the format \code{module$function}}

\item{help_type}{character string specifying the output format; currently,
only \code{'text'} is supported}
}
\description{
\code{module_help} displays help on a module’s objects and functions in much
the same way \code{\link[utils]{help}} does for package contents.
}
\note{
Help is only available if \code{\link{import}} loaded the help stored
in the module file(s). By default, this happens only in interactive sessions.
}
\examples{
\dontrun{
mod = import('mod')
module_help(mod$func)
}
\dontrun{
?mod$func
}
\dontrun{
help(mod$func)
}
}

