% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{info}
\alias{info}
\alias{mod_info}
\alias{pkg_info}
\title{Information about a physical module or package}
\usage{
mod_info(spec, source_path)

pkg_info(spec)
}
\arguments{
\item{source_path}{character string full path to the physical module
location.}

\item{mod_spec}{a \code{mod_spec}.}

\item{pkg_spec}{a \code{pkg_spec}}
}
\value{
\code{mod_info} and \code{pkg_info} return a structure representing
the module/package information for the given specification/source location.
}
\description{
A \code{mod_info} represents an existing, installed module and its runtime
physical location (usually in the file system).
}
\keyword{internal}
