% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.r
\name{is_S3_user_generic}
\alias{is_S3_user_generic}
\title{Check whether a function given by name is a user-defined generic}
\usage{
is_S3_user_generic(function_name, envir = parent.frame())
}
\arguments{
\item{function_name}{function name as character string}

\item{envir}{the environment this function is invoked from}
}
\description{
A user-defined generic is any function which, at some point, calls
}

