% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.r
\name{make_S3_methods_known}
\alias{make_S3_methods_known}
\title{Find and register S3 methods inside a module}
\usage{
make_S3_methods_known(module)
}
\arguments{
\item{module}{the module object for which to register S3 methods}
}
\description{
Find and register S3 methods inside a module
}

