% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file.r
\name{module_file}
\alias{module_file}
\title{Find the full file names of files in modules}
\usage{
module_file(..., module = parent.frame(), mustWork = FALSE)
}
\arguments{
\item{...}{character vectors of files or subdirectories inside a module; if
none is given, return the root directory of the module}

\item{module}{a module environment (default: current module)}

\item{mustWork}{logical; if \code{TRUE}, an error is raised if the given
files do not match existing files.}
}
\value{
A character vector containing the absolute paths to the files
 specified in \code{...}, or an empty string, \code{''}, if no file was
 found (unless \code{mustWork = TRUE} was specified).
}
\description{
Find the full file names of files in modules
}
\note{
If called from outside a module, the current working directory is used.

This function is similar to \code{system.file} for packages. It is provided
as a separate function rather than overriding \code{system.file} because that
would cause ambiguity when a module and a package share the same name.
}
\seealso{
\code{base::system.file}
}

