\docType{package}
\name{modules}
\alias{modules}
\alias{modules-package}
\title{An alternative module system for R}
\description{
Use \code{module = import(module)} to import a module for
usage. Fully qualified names are supported for nested
modules, reminiscent of Python’s module mechanism.
}
\seealso{
\code{import}

\code{module_name}
}

