% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.r
\name{register_S3_method}
\alias{register_S3_method}
\title{Register an S3 method for a given generic and class.}
\usage{
register_S3_method(name, class, method)
}
\arguments{
\item{name}{the name of the generic as a character string}

\item{class}{the class name}

\item{method}{the method to register}
}
\description{
Register an S3 method for a given generic and class.
}
\details{
Methods for generics defined in the same module do not need to be
registered explicitly, and indeed \emph{should not} be registered. However,
if the user wants to add a method for a known generic (e.g.
\code{\link{print}}), then this needs to be made known explicitly.
}
\note{
\strong{Do not} call \code{\link{registerS3method}} inside a module.
Use \code{register_S3_method} instead – this is important for the module’s
own book-keeping.
}
\examples{
\dontrun{
# In module a:
print.my_class = function (x) {
    cat(sprintf('My class with field \%s\\n', x$field))
    invisible(x)
}

register_S3_method('print', 'my_class', print.my_class)

# Globally:
a = import('a')
obj = structure(list(field = 42), class = 'my_class')
obj # calls `print`, with output "My class with field 42"
}
}

