% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{reload}
\alias{reload}
\title{Reload a given module}
\usage{
reload(module)
}
\arguments{
\item{module}{reference to the module which should be unloaded}
}
\description{
Remove the loaded module from the cache, forcing a reload. The newly reloaded
module is assigned to the module reference in the calling scope.
}
\note{
Any other references to the loaded modules remain unchanged, and will
still work. Reloading modules is primarily useful for testing during
development, and should not be used in production code.

\code{reload} comes with a few restrictions. It attempts to re-attach itself
in parts or whole if it was previously attached in parts or whole. This only
works if it is called in the same scope as the original \code{import}.
}
\seealso{
\code{import}

\code{unload}
}

