% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cache.r
\name{set_script_path}
\alias{set_script_path}
\title{Set the base path of the script.}
\usage{
set_script_path(path)
}
\arguments{
\item{path}{character string containing the relative or absolute path, or
\code{NULL} to reset the path}
}
\description{
Set the base path of the script.
}
\details{
\emph{modules} needs to know the base path of the topmost calling R script
to find relative import locations. In most cases, it can figure the path out
automatically. However, in some cases third party packages load files in such
a way that \emph{modules} cannot find out the correct path of the script any
more. \code{set_script_path} can be used in these cases to set the script
path manually.
}

