% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{unload}
\alias{unload}
\title{Unload a given module}
\usage{
unload(module)
}
\arguments{
\item{module}{reference to the module which should be unloaded}
}
\description{
Unset the module variable that is being passed as a parameter, and remove the
loaded module from cache.
}
\note{
Any other references to the loaded modules remain unchanged, and will
still work. However, subsequently importing the module again will reload its
source files, which would not have happened without \code{unload}.
Unloading modules is primarily useful for testing during development, and
should not be used in production code.

\code{unload} comes with a few restrictions. It attempts to detach itself
if it was previously attached. This only works if it is called in the same
scope as the original \code{import}.
}
\seealso{
\code{import}

\code{reload}
}

